% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.data.frame.lfq_data}
\alias{as.data.frame.lfq_data}
\title{Convert lfq_data to long-format tibble}
\usage{
\method{as.data.frame}{lfq_data}(x, ...)
}
\arguments{
\item{x}{An \link{lfq_data} object.}

\item{...}{Ignored.}
}
\value{
A tibble with all columns.
}
\description{
Convert lfq_data to long-format tibble
}
\examples{
data(sarscov2_us_2022)
x <- lfq_data(sarscov2_us_2022, lineage = variant,
              date = date, count = count, total = total)
as.data.frame(x)

}
