% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as_lfq_data}
\alias{as_lfq_data}
\alias{as_lfq_data.lfq_data}
\alias{as_lfq_data.data.frame}
\title{Coerce to lfq_data}
\usage{
as_lfq_data(x, ...)

\method{as_lfq_data}{lfq_data}(x, ...)

\method{as_lfq_data}{data.frame}(x, ...)
}
\arguments{
\item{x}{An object to coerce.}

\item{...}{Additional arguments passed to methods. For the
\code{data.frame} method, these are passed to \code{\link[=lfq_data]{lfq_data()}}.}
}
\value{
An \link{lfq_data} object.

An \link{lfq_data} object.

An \link{lfq_data} object.
}
\description{
Generic function to convert various data formats into \link{lfq_data}
objects. Methods can be defined for specific input classes to enable
seamless interoperability with other genomic surveillance packages.
}
\examples{
df <- data.frame(
  date    = rep(as.Date("2024-01-01") + c(0, 7), each = 2),
  lineage = rep(c("A", "B"), 2),
  count   = c(80, 20, 60, 40)
)
x <- as_lfq_data(df, lineage = lineage, date = date, count = count)
x

}
\seealso{
\code{\link[=lfq_data]{lfq_data()}} for the primary constructor.
}
