% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{autoplot.lfq_forecast}
\alias{autoplot.lfq_forecast}
\title{Plot a lineage frequency forecast}
\usage{
\method{autoplot}{lfq_forecast}(object, ...)
}
\arguments{
\item{object}{An \link{lfq_forecast} object.}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Plot a lineage frequency forecast
}
\examples{
sim <- simulate_dynamics(n_lineages = 3,
  advantages = c("A" = 1.2, "B" = 0.8), seed = 1)
fit <- fit_model(sim)
fc <- forecast(fit, horizon = 14)
autoplot(fc)

}
