% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecast lineage frequencies (generic)}
\usage{
forecast(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A forecast object.
}
\description{
Forecast lineage frequencies (generic)
}
