% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq-data.R
\name{is_lfq_data}
\alias{is_lfq_data}
\title{Test if an object is an lfq_data object}
\usage{
is_lfq_data(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical scalar.
}
\description{
Test if an object is an lfq_data object
}
\examples{
d <- data.frame(date = Sys.Date(), lineage = "A", count = 10)
x <- lfq_data(d, lineage = lineage, date = date, count = count)
is_lfq_data(x)
is_lfq_data(d)

}
