% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-api.R
\name{lfq_advantage}
\alias{lfq_advantage}
\title{Pipe-friendly growth advantage extraction}
\usage{
lfq_advantage(fit, type = "relative_Rt", generation_time = NULL, ...)
}
\arguments{
\item{fit}{An \code{lfq_fit} object.}

\item{type}{Output type. Default \code{"relative_Rt"}.}

\item{generation_time}{Mean generation time in days.}

\item{...}{Passed to \code{\link[=growth_advantage]{growth_advantage()}}.}
}
\value{
A tibble of growth advantages.
}
\description{
Pipe-friendly growth advantage extraction
}
