% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{lfq_engines}
\alias{lfq_engines}
\title{List available modeling engines}
\usage{
lfq_engines()
}
\value{
A tibble with columns: \code{engine}, \code{type}, \code{time_varying},
\code{available}, \code{description}.
}
\description{
Returns information about all modeling engines available in
lineagefreq. Core engines (mlr, hier_mlr, piantham) are always
available. Bayesian engines (fga, garw) require 'CmdStan'.
}
\examples{
lfq_engines()

}
