% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-api.R
\name{lfq_fit}
\alias{lfq_fit}
\title{Pipe-friendly model fitting}
\usage{
lfq_fit(data, engine = "mlr", ...)
}
\arguments{
\item{data}{An \link{lfq_data} object.}

\item{engine}{Engine name. Default \code{"mlr"}.}

\item{...}{Passed to \code{\link[=fit_model]{fit_model()}}.}
}
\value{
An \code{lfq_fit} object.
}
\description{
Enables tidyverse-style chaining:
\code{data |> lfq_fit("mlr") |> lfq_forecast(28) |> lfq_score()}
}
\examples{
data(sarscov2_us_2022)
sarscov2_us_2022 |>
  lfq_data(lineage = variant, date = date, count = count, total = total) |>
  lfq_fit("mlr") |>
  lfq_advantage(generation_time = 5)

}
