% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-api.R
\name{lfq_forecast}
\alias{lfq_forecast}
\title{Pipe-friendly forecasting}
\usage{
lfq_forecast(fit, horizon = 28L, ...)
}
\arguments{
\item{fit}{An \code{lfq_fit} object.}

\item{horizon}{Forecast horizon in days. Default 28.}

\item{...}{Passed to \code{\link[=forecast]{forecast()}}.}
}
\value{
An \code{lfq_forecast} object.
}
\description{
Pipe-friendly forecasting
}
