% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validation.R
\name{lfq_stan_available}
\alias{lfq_stan_available}
\title{Check if CmdStan backend is available}
\usage{
lfq_stan_available()
}
\value{
Logical scalar.
}
\description{
Returns \code{TRUE} if CmdStanR and CmdStan are installed.
Bayesian engines require this.
}
\examples{
lfq_stan_available()

}
