% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineagefreq-package.R
\docType{package}
\name{lineagefreq-package}
\alias{lineagefreq}
\alias{lineagefreq-package}
\title{lineagefreq: Lineage Frequency Dynamics from Genomic
Surveillance Counts}
\description{
Models pathogen lineage frequency dynamics from genomic
surveillance count data. Provides a unified \code{fit_model()}
interface with multiple engines (MLR, hierarchical MLR,
Piantham), rolling-origin backtesting via \code{backtest()},
standardized scoring via \code{score_forecasts()}, emergence
detection, and sequencing power analysis.
}
\section{Quick start}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{x <- lfq_data(my_counts, lineage = lineage, date = date, count = n)
fit <- fit_model(x, engine = "mlr")
growth_advantage(fit, generation_time = 5)
}\if{html}{\out{</div>}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CuiweiG/lineagefreq}
  \item Report bugs at \url{https://github.com/CuiweiG/lineagefreq/issues}
}

}
\author{
\strong{Maintainer}: Cuiwei Gao \email{48gaocuiwei@gmail.com}

}
\keyword{internal}
