% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coef.R
\name{logLik.lfq_fit}
\alias{logLik.lfq_fit}
\title{Extract log-likelihood from an lfq_fit object}
\usage{
\method{logLik}{lfq_fit}(object, ...)
}
\arguments{
\item{object}{An \link{lfq_fit} object.}

\item{...}{Ignored.}
}
\value{
An object of class \code{logLik}.
}
\description{
Extract log-likelihood from an lfq_fit object
}
\examples{
sim <- simulate_dynamics(
  n_lineages = 3, advantages = c("A" = 1.2, "B" = 0.8),
  n_timepoints = 12, seed = 1
)
fit <- fit_model(sim, engine = "mlr")
logLik(fit)

}
