% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coef.R
\name{nobs.lfq_fit}
\alias{nobs.lfq_fit}
\title{Extract number of observations from an lfq_fit object}
\usage{
\method{nobs}{lfq_fit}(object, ...)
}
\arguments{
\item{object}{An \link{lfq_fit} object.}

\item{...}{Ignored.}
}
\value{
Integer scalar (total sequence count used in fitting).
}
\description{
Extract number of observations from an lfq_fit object
}
\examples{
sim <- simulate_dynamics(
  n_lineages = 3, advantages = c("A" = 1.2, "B" = 0.8),
  n_timepoints = 12, seed = 1
)
fit <- fit_model(sim, engine = "mlr")
nobs(fit)

}
