% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive-design.R
\name{plot.adaptive_allocation}
\alias{plot.adaptive_allocation}
\title{Plot adaptive allocation}
\usage{
\method{plot}{adaptive_allocation}(x, ...)
}
\arguments{
\item{x}{An \code{adaptive_allocation} object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot adaptive allocation
}
