% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveillance-value.R
\name{plot.evoi}
\alias{plot.evoi}
\title{Plot EVOI curve}
\usage{
\method{plot}{evoi}(x, ...)
}
\arguments{
\item{x}{An \code{evoi} object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot EVOI curve
}
