% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness-decomposition.R
\name{plot.fitness_decomposition}
\alias{plot.fitness_decomposition}
\title{Plot fitness decomposition}
\usage{
\method{plot}{fitness_decomposition}(x, ...)
}
\arguments{
\item{x}{A \code{fitness_decomposition} object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Stacked bar plot showing the partition of growth advantage into
intrinsic transmissibility and immune escape components.
}
