% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune-landscape.R
\name{plot.immune_landscape}
\alias{plot.immune_landscape}
\title{Plot population immunity landscape}
\usage{
\method{plot}{immune_landscape}(x, ...)
}
\arguments{
\item{x}{An \code{immune_landscape} object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot population immunity landscape
}
