% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness-decomposition.R
\name{tidy.fitness_decomposition}
\alias{tidy.fitness_decomposition}
\title{Tidy a fitness decomposition}
\usage{
tidy.fitness_decomposition(x, ...)
}
\arguments{
\item{x}{A \code{fitness_decomposition} object.}

\item{...}{Unused.}
}
\value{
A tibble with the decomposition results.
}
\description{
Tidy a fitness decomposition
}
